requirements.txt
Flask==2.3.3
speedtest-cli==2.1.3